Set up tox-conda/pytest
- Set build-backend so isolated pip install works
- Add pychop as testing dependency
- Configure tox-conda
- Create __init__.py files under tests/ so that all tests are
  found by pytest

We need to use conda to install Mantid and check results against Abins
functions. Tox-conda is stuck on tox v3; there is a branch to support
v4 but it's not clear if/when this will be finished. But we also might
be able to drop the conda component of these tests once we are
satisfied with validation against Mantid: it can be replaced with
tests against reference values.

Perhaps we don't need the test dependencies at all in pyproject.toml,
now this is all defined in tox.ini. But they could be useful for
running tests without tox if desired.